Make it viewer.loader for now
